Move the IDE component to the containers folder
