+++ game_frontend/src/containers/IDE/index.test.js
@@ -2,7 +2,7 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 import toJson from 'enzyme-to-json'
-import IDE from 'components/IDE'
+import IDE from 'containers/IDE'
 
 describe('<IDE />', () => {
   it('renders correctly', () => {
